+++ public/admin/invoices/index.php
@@ -30,6 +30,13 @@
 			echo $adminUIRender->renderAdminSideBar('../../');
 		?>
 
+		<?php
+            require_once '../../../lib/render/ui/popupsHandler.php';
+            $popupsHandler = new popupsHandler('mainPopupsHandler', ["rootPathPrefix" => '../../', 'class' => 'styledText spacedText defaultMainShadows']);
+            $popupsHandler->render();
+            echo $popupsHandler->output;
+        ?>
+
 		<div class="cmsMainContentWrapper styledText textColorThemeInvoices">
 			<div style="margin-left: 2em; margin-right: 2em;" class="spacedText">
 				<h1>Invoices</h1>